pyproject.toml
[tool.poetry]
name = "ea-see"
version = "0.1.0"
description = ""
authors = ["Da-Emp66 <asherbean10@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
opencv-python = "^4.10.0.84"
torch = "^2.4.1"
python-dotenv = "^1.0.1"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
torchfile = "^0.1.0"
tensorflow = "^2.17.0"
torchvision = "^0.20.0"
gdown = "^5.2.0"
tensorrt = "^10.5.0"
ordered-set = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
